extract new code to bits.appeal.ts
